feat: add the ability to change themes

+++ src/shared/db/graphql/resolvers.ts
@@ -28,10 +28,17 @@ export const resolvers = {
             await dbConnect();
             const listsIds = await List.find({ boardId: id }).distinct('_id');
 
-            await Task.deleteMany({ listId: { $in: listsIds } });
-            await Label.deleteMany({ boardId: id });
-            await List.deleteMany({ boardId: id });
-            await Board.deleteOne({ _id: id });
+            if (!listsIds.length) {
+                await Board.deleteOne({ _id: id });
+                return id;
+            }
+
+            await Promise.all([
+                Task.deleteMany({ listId: { $in: listsIds } }),
+                Label.deleteMany({ boardId: id }),
+                List.deleteMany({ boardId: id }),
+                Board.deleteOne({ _id: id }),
+            ]);
 
             return id;
         },
@@ -75,6 +82,33 @@ export const resolvers = {
         createBoardsGroup: async (_: any, { name, order }: { name: string; order: number }) => {
             await dbConnect();
             return BoardsGroup.create({ name, order });
+        },
+        deleteBoardsGroup: async (_: any, { id }: { id: string }) => {
+            await dbConnect();
+
+            const boardsIds = await Board.find({ groupId: id }).distinct('_id');
+            if (!boardsIds.length) {
+                await BoardsGroup.deleteOne({ _id: id });
+                return id;
+            }
+
+            const listsIds = await List.find({ boardId: { $in: boardsIds } }).distinct('_id');
+
+            await Promise.all([
+                Task.deleteMany({ listId: { $in: listsIds } }),
+                Label.deleteMany({ boardId: { $in: boardsIds } }),
+                List.deleteMany({ boardId: { $in: boardsIds } }),
+                Board.deleteMany({ groupId: id }),
+                BoardsGroup.deleteOne({ _id: id }),
+            ]);
+
+            return id;
+        },
+        updateBoardsGroup: async (_: any, { id, name }: { id: string; name: string }) => {
+            await dbConnect();
+            await BoardsGroup.updateOne({ _id: id }, { $set: { name } });
+
+            return BoardsGroup.findById(id);
         },
     },
     BoardsGroup: {
